# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Music Catalog API",
        "description": "API для управления исполнителями, альбомами и песнями.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Localhost API server"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Localhost API server"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Localhost API server"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Localhost API server"
        }
    ],
    "paths": {
        "/api/albums": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Получить список всех альбомов",
                "operationId": "735609843dcfb89ee4e12e340eddca25",
                "responses": {
                    "200": {
                        "description": "Успешное выполнение"
                    }
                }
            },
            "post": {
                "tags": [
                    "Albums"
                ],
                "summary": "Создать новый альбом",
                "operationId": "2252aa8fd63b7d2fa1203c1735943d46",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "artist_id",
                                    "release_year"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "A Head Full of Dreams"
                                    },
                                    "artist_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "release_year": {
                                        "type": "integer",
                                        "example": 2015
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Альбом создан"
                    }
                }
            }
        },
        "/api/albums/{id}": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Получить информацию об альбоме",
                "operationId": "71e9c4179582355e89419e6f6994c418",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID альбома",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об альбоме"
                    }
                }
            },
            "put": {
                "tags": [
                    "Albums"
                ],
                "summary": "Обновить информацию об альбоме",
                "operationId": "1300ec85cc590b1ae1468c9bf8db122d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID альбома",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "artist_id",
                                    "release_year"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Mylo Xyloto"
                                    },
                                    "artist_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "release_year": {
                                        "type": "integer",
                                        "example": 2011
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Альбом обновлен"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Albums"
                ],
                "summary": "Удалить альбом",
                "operationId": "cabfc7c152115607d310827d46a82d5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID альбома",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Альбом удален"
                    }
                }
            }
        },
        "/api/albums/{album}/songs": {
            "get": {
                "tags": [
                    "Album Songs"
                ],
                "summary": "Получить список песен в альбоме",
                "operationId": "4bfc954afef493a8e4225295cca0d03a",
                "parameters": [
                    {
                        "name": "album",
                        "in": "path",
                        "description": "ID альбома",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список песен в альбоме",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Song Title"
                                            },
                                            "track_number": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Альбом не найден"
                    }
                }
            },
            "post": {
                "tags": [
                    "Album Songs"
                ],
                "summary": "Добавить песню в альбом",
                "operationId": "48ae422672a44fbc2da68e7c3b07f0ae",
                "parameters": [
                    {
                        "name": "album",
                        "in": "path",
                        "description": "ID альбома",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "song_id",
                                    "track_number"
                                ],
                                "properties": {
                                    "song_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "track_number": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Песня добавлена в альбом"
                    },
                    "404": {
                        "description": "Альбом или песня не найдены"
                    }
                }
            }
        },
        "/api/albums/{album}/songs/{song}": {
            "delete": {
                "tags": [
                    "Album Songs"
                ],
                "summary": "Удалить песню из альбома",
                "operationId": "950426a912db2383f8d68ce10762162f",
                "parameters": [
                    {
                        "name": "album",
                        "in": "path",
                        "description": "ID альбома",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "song",
                        "in": "path",
                        "description": "ID песни",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Песня удалена из альбома"
                    },
                    "404": {
                        "description": "Альбом или песня не найдены"
                    }
                }
            }
        },
...
    },
...
}